# JIT: prolog zeroing accounting changes