@AndyAyersMS 2019-03-27
Follow-up to comments at tail end of #23403.

Upshot is that we will now use multiple stores to zero up to 32 bytes on a 64 bit platform instead of a loop or `rep stos`.

Still trying to get a fair evaluation of the perf impact. On a simple test like:
```C#
    [MethodImpl(MethodImplOptions.NoInlining)]
    private void Dummy(Span<byte> a, Span<byte> b) { }

    [MethodImpl(MethodImplOptions.NoInlining)]
    public void StackZero() { Span<byte> a, b; a = bytes; b = bytes; Dummy(a, b); }
```
it is substantial, about a 2x improvement. Wider tests (using the performance repo) show a lot of potential good and bad results (more good than bad) and I am working through some of the cases and will summarize when I have something interesting to say.

Code size is generally improved:
```
Total bytes of diff: -20050 (-0.05% of base)
    diff is an improvement.

Top file regressions by size (bytes):
         122 : CommandLine.dasm (0.03% of base)
          45 : NuGet.Packaging.dasm (0.03% of base)
          19 : System.ComponentModel.Annotations.dasm (0.04% of base)
          11 : Microsoft.DotNet.InternalAbstractions.dasm (0.22% of base)
           6 : NuGet.Packaging.Core.Types.dasm (0.04% of base)

Top file improvements by size (bytes):
       -3849 : System.Private.CoreLib.dasm (-0.09% of base)
       -3071 : Microsoft.CodeAnalysis.VisualBasic.dasm (-0.05% of base)
       -2137 : Microsoft.CodeAnalysis.CSharp.dasm (-0.05% of base)
       -1266 : Microsoft.CodeAnalysis.dasm (-0.07% of base)
        -745 : System.Collections.Immutable.dasm (-0.07% of base)

98 total files with size differences (84 improved, 14 regressed), 31 unchanged.

Top method regressions by size (bytes):
          66 ( 2.10% of base) : System.Linq.Parallel.dasm - NullableDecimalAverageAggregationOperatorEnumerator`1:MoveNextCore(byref):bool:this (7 methods)
          66 ( 2.18% of base) : System.Linq.Parallel.dasm - NullableDecimalSumAggregationOperatorEnumerator`1:MoveNextCore(byref):bool:this (7 methods)
          14 ( 0.30% of base) : CommandLine.dasm - Result`2:ToString():ref:this (7 methods)
          14 ( 0.41% of base) : Microsoft.CodeAnalysis.VisualBasic.dasm - MethodToClassRewriter`1:VisitMethodSymbol(ref):ref:this (7 methods)
          14 ( 0.23% of base) : Microsoft.CodeAnalysis.VisualBasic.dasm - SynthesizedWrapperMethod:.ctor(ref,ref,ref,ref):this (7 methods)

Top method improvements by size (bytes):
         -96 (-1.01% of base) : System.Private.CoreLib.dasm - MemoryExtensions:TrimEnd(struct,struct):struct (29 methods)
         -91 (-9.35% of base) : Microsoft.DotNet.Cli.Utils.dasm - Result`1:AsValue(ubyte):struct:this (7 methods)
         -91 (-9.29% of base) : Microsoft.DotNet.Cli.Utils.dasm - Result`1:AsValue(short):struct:this (7 methods)
         -91 (-9.63% of base) : Microsoft.DotNet.Cli.Utils.dasm - Result`1:AsValue(int):struct:this (7 methods)
         -91 (-9.56% of base) : Microsoft.DotNet.Cli.Utils.dasm - Result`1:AsValue(double):struct:this (7 methods)

Top method regressions by size (percentage):
          11 ( 3.90% of base) : System.Collections.dasm - TreeSubSet:get_MinInternal():struct:this
          11 ( 3.90% of base) : System.Collections.dasm - TreeSubSet:get_MaxInternal():struct:this
           6 ( 3.49% of base) : Microsoft.DotNet.Cli.Utils.dasm - DebugHelper:WaitForDebugger()
           5 ( 3.11% of base) : Microsoft.DotNet.InternalAbstractions.dasm - RuntimeEnvironment:GetRIDArch():ref
          11 ( 2.65% of base) : System.Net.Security.dasm - SslStreamPal:AcquireCredentialsHandle(ref,int,int,bool):ref

Top method improvements by size (percentage):
         -17 (-24.29% of base) : System.Security.Cryptography.Algorithms.dasm - AlgorithmIdentifierAsn:HasNullEquivalentParameters():bool:this
         -17 (-24.29% of base) : System.Security.Cryptography.Cng.dasm - AlgorithmIdentifierAsn:HasNullEquivalentParameters():bool:this
         -17 (-23.61% of base) : System.Private.CoreLib.dasm - TimeSpanParse:TryParseTimeSpanConstant(struct,byref):bool
         -12 (-17.65% of base) : System.Data.Common.dasm - SqlString:ToSqlBoolean():struct:this
         -12 (-17.65% of base) : System.Data.Common.dasm - SqlString:ToSqlByte():struct:this

4516 total methods with size differences (3603 improved, 913 regressed), 195607 unchanged.
```
Regressions are from cases where we have large offsets that require multibyte encoding. Improvements are from not needing to preserve as many registers in some cases.

Not considering xmm init just yet as that is a more involved change and post 3.0 we might want to unify or at least share strategy for prolog init with `initBlk`.

Also not (yet) considering trying to order the stores by increasing offset or be smarter about non initlocals cases where we have structs with non-gc fields.

cc @dotnet/jit-contrib