@AndyAyersMS 2019-03-27
Most significant improvements (from performance repo CoreCLR tests). Filtered out a few by hand that did not seem credible.

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) |
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:|
| System.Memory.MemoryMarshal\<Int32>.TryGetArray                                   |      2.78 |             9.63 |             3.47 |
| System.Memory.MemoryMarshal\<Byte>.TryGetArray                                    |      2.21 |             9.81 |             4.44 |
| System.Tests.Perf_String.Insert(s1: "Test", i: 2, s2: " Test")                   |      1.92 |            23.70 |            12.36 |
| System.Collections.IndexerSet<String>.Array(Size: 512)                           |      1.64 |           278.27 |           169.46 |
| System.Tests.Perf_String.Insert(s1: "dzsdzsDDZSDZSDZSddsz", i: 7, s2: "Test")    |      1.63 |            22.64 |            13.90 |
| System.Collections.CopyTo\<Int32>.ReadOnlyMemory(Size: 2048)                      |      1.58 |           157.75 |            99.91 |

Still sorting through potential regressions...