@AndyAyersMS 2019-03-28
Finished looking at a bunch of potential regressions. None of them seem to hold up under scrutiny -- many look the same or were improvements on rerun, and the remaining few that still looked worse did not have codegen diffs (one seemed to be a 32 bit alignment issue, another had GC kicking in). Will list them here in the interest of full disclosure.

| Slower                                                                     | diff/base | Base Median (ns) | Diff Median (ns) |
| -------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:|
| Devirtualization.EqualityComparer.ValueTupleCompareWrapped                 |      3.01 |             4.76 |            14.33 |
| System.Memory.Constructors<Byte>.ReadOnlySpanFromArray                     |      2.51 |             0.23 |             0.57 |
| System.Collections.Sort<Int32>.Array(Size: 512)                            |      1.84 |          3742.65 |          6902.07 |
| System.Collections.IndexerSet<Int32>.Span(Size: 512)                       |      1.58 |           168.02 |           265.51 |
| System.Collections.CopyTo<Int32>.Span(Size: 2048)                          |      1.55 |            97.84 |           151.85 |
| System.Collections.CopyTo<Int32>.ReadOnlySpan(Size: 2048)                  |      1.53 |            99.31 |           152.07 |
| System.Tests.Perf_String.IndexerCheckLengthHoisting                        |      1.44 |            25.86 |            37.18 |
| PerfLabTests.LowLevelPerf.GenericGenericMethod                             |      1.40 |        127311.62 |        178717.48 |
| System.Collections.IndexerSetReverse<Int32>.Span(Size: 512)                |      1.31 |           203.07 |           265.72 |
| System.MathBenchmarks.Double.Atan2                                         |      1.23 |         89569.64 |        109864.78 |